requirements.txt
amqp==5.0.9
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Deprecated==1.2.13
h11==0.12.0
idna==3.3
kombu==5.2.2
outcome==1.1.0
packaging==21.3
prompt-toolkit==3.0.24
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
redis==4.1.0
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
wsproto==1.0.0
